backend/requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Security
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# ML/AI (Optional)
scikit-learn==1.3.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
